Moved enums to first base usage file
