+++ TownOfUs/Modifiers/Game/TouGameModifier.cs
@@ -18,73 +18,3 @@ public abstract class TouGameModifier : TouBaseGameModifier
         return !role.Player.GetModifierComponent().HasModifier<TouGameModifier>(true, x => x.PreventsOtherModifiers) && role is not SpectatorRole;
     }
 }
-
-public enum ModifierFaction
-{
-    Alliance,
-    Universal,
-    Crewmate,
-    Neutral,
-    Impostor,
-    CrewmateAlliance,
-    CrewmateUtility,
-    CrewmateVisibility,
-    CrewmatePostmortem,
-    CrewmatePassive,
-    NeutralAlliance,
-    NeutralUtility,
-    NeutralVisibility,
-    NeutralPostmortem,
-    NeutralPassive,
-    ImpostorAlliance,
-    ImpostorUtility,
-    ImpostorVisibility,
-    ImpostorPostmortem,
-    ImpostorPassive,
-    UniversalUtility,
-    UniversalVisibility,
-    UniversalPostmortem,
-    UniversalPassive,
-    AssailantUtility,
-    AssailantVisibility,
-    AssailantPostmortem,
-    AssailantPassive,
-    NonCrewmate,
-    NonCrewUtility,
-    NonCrewVisibility,
-    NonCrewPostmortem,
-    NonCrewPassive,
-    NonNeutral,
-    NonNeutUtility,
-    NonNeutVisibility,
-    NonNeutPostmortem,
-    NonNeutPassive,
-    NonImpostor,
-    NonImpUtility,
-    NonImpVisibility,
-    NonImpPostmortem,
-    NonImpPassive,
-    HiderUtility,
-    HiderVisibility,
-    HiderPostmortem,
-    HiderPassive,
-    SeekerUtility,
-    SeekerVisibility,
-    SeekerPostmortem,
-    SeekerPassive,
-    External,
-    Other
-}
-
-public enum AlliedFaction
-{
-    Crewmate,
-    CrewmateKiller,
-    Neutral,
-    NeutralKiller,
-    Impostor,
-    Lover,
-    Recruit,
-    RoleSpecific,
-    Other,
-}
