+++ TownOfUs/Modifiers/HnsGame/Crewmate/HnsTransporterModifier.cs
@@ -5,7 +5,6 @@ using MiraAPI.Utilities.Assets;
 using Reactor.Networking.Attributes;
 using Reactor.Utilities.Extensions;
 using TownOfUs.Interfaces;
-using TownOfUs.Modifiers.Game;
 using TownOfUs.Modules;
 using TownOfUs.Options.Modifiers;
 using UnityEngine;
